more refactoring of logging
