+++ ZipUtility/LogApi/appsettings.json
@@ -12,7 +12,8 @@
         "Microsoft.Hosting.Lifetime": "Information",
         "LogApi": "Trace"
       },
-      "Folder": "./bin/Logs"
+      "Folder": "./bin/Logs",
+      "FileExtension": ".log"
     },
     "Console": {
       "LogLevel": {
